# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passwordless Multi-Auth API",
    "description": "Passwordless Multi-Auth Platform API - Passwordless Authentication\n\nThis API provides comprehensive endpoints for:\n- MagicLink-based passwordless authentication\n- User management and security operations\n- Role and authority management\n- Push notifications\n- TOTP/MFA verification\n\n## Authentication Flow\n1. Call `POST /api/auth/email-magiclink/send` with user email\n2. User opens the emailed link in a browser: `GET /api/auth/verify?token={token}`\n3. The server redirects (302) to the app scheme `passwordless://auth/callback?code={code}` (or `?status=totp_required` when MFA is enabled, `?status=error&reason=...` on failure)\n4. Trade the one-time code for a token pair via `POST /api/auth/exchange` (if TOTP was required, call `POST /api/auth/totp/verify` instead)\n5. Use the returned JWT in the Authorization header: `Bearer {token}`\n6. Renew with `POST /api/auth/refresh` (single-use rotation); revoke with `POST /api/auth/logout`\n\n## Flutter Integration\nFor Flutter apps, use the `/api/auth/email-magiclink/send` endpoint\nfor mobile authentication. The MagicLink can be handled via\ndeep linking in your Flutter app.\n",
    "contact": {
      "name": "Passwordless Multi-Auth Team",
      "email": "api@example.com"
    },
    "license": {
      "name": "Private",
      "url": "https://github.com/ersincivi/spring-boot-passwordless-multi-auth"
    },
    "version": "1.0.0"
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/settings/qr-code": {
      "get": {
        "tags": [
          "settings-controller"
        ],
        "operationId": "getQrCodeImage",
        "parameters": [
          {
            "name": "totpUri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter JWT token obtained from MagicLink authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}